Quick mod to fix incorrect change to v2.x.x.
Original mod intended to make this package a v2.x.x backward-incompatible
version.  The module-izing was not done correctly resulting in
applications using this module to fail to compile.

* Partially resolves CASMHMS-3088

Tested on:

* Build machine

The same testing was done as with the original mod.  All tests
passed correctly.

This particular mod is low risk as it fixes a build problem.